Feat: Better naming for CollectDependencies
